% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_to_global}
\alias{assign_to_global}
\title{function loading results in global environment}
\usage{
assign_to_global(pos = 1, what, value)
}
\arguments{
\item{pos}{defaults to 1 which equals an assingment to global environment}

\item{what}{variable to assign to}

\item{value}{value to assign}
}
\description{
function loading results in global environment
}
