% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{func1}
\alias{func1}
\title{An internal function to compute m and gamma based on 
continuous-time model (Yashin et. al., 2007)}
\usage{
func1(tt, y, a, f1, Q, f, b, theta)
}
\arguments{
\item{tt}{tt - time}

\item{y}{y}

\item{a}{a (see Yashin et. al, 2007)}

\item{f1}{f1 (see Yashin et. al, 2007)}

\item{Q}{Q (see Yashin et. al, 2007)}

\item{f}{f (see Yashin et. al, 2007)}

\item{b}{b (see Yashin et. al, 2007)}

\item{theta}{theta}
}
\value{
list(m, gamma) Next values of m and gamma (see Yashin et. al, 2007)
}
\description{
An internal function to compute m and gamma based on 
continuous-time model (Yashin et. al., 2007)
}
