% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get.column.index}
\alias{get.column.index}
\title{An internal function to obtain column index by its name}
\usage{
get.column.index(x, col.name)
}
\arguments{
\item{x}{Dataset}

\item{col.name}{Column name}
}
\value{
column index(es) in the provided dataset
}
\description{
An internal function to obtain column index by its name
}
