% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{getNextY.cont}
\alias{getNextY.cont}
\title{An internal function to compute next Y based on 
continous-time model (Yashin et. al., 2007)}
\usage{
getNextY.cont(y1, t1, t2, a, f1, Q, f, b, theta)
}
\arguments{
\item{y1}{y1}

\item{t1}{t1}

\item{t2}{t2}

\item{a}{a (see Yashin et. al, 2007)}

\item{f1}{f1 (see Yashin et. al, 2007)}

\item{Q}{Q (see Yashin et. al, 2007)}

\item{f}{f (see Yashin et. al, 2007)}

\item{b}{b (see Yashin et. al, 2007)}

\item{theta}{theta (see Yashin et. al, 2007)}
}
\value{
y.next Next value of Y
}
\description{
An internal function to compute next Y based on 
continous-time model (Yashin et. al., 2007)
}
