% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{getNextY.cont2}
\alias{getNextY.cont2}
\title{An internal function to compute next value of physiological variable Y}
\usage{
getNextY.cont2(y1, t1, t2, b, a, f1)
}
\arguments{
\item{y1}{y1}

\item{t1}{t1}

\item{t2}{t2}

\item{b}{b (see Yashin et. al, 2007)}

\item{a}{a (see Yashin et. al, 2007)}

\item{f1}{f1 (see Yashin et. al, 2007)}
}
\value{
y.next Next value of y
}
\description{
An internal function to compute next value of physiological variable Y
}
