% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{getNextY.discr.m}
\alias{getNextY.discr.m}
\title{An internal function to compute next m based on dicrete-time model}
\usage{
getNextY.discr.m(y1, u, R)
}
\arguments{
\item{y1}{y1}

\item{u}{u}

\item{R}{R}
}
\value{
m Next value of m (see Yashin et. al, 2007)
}
\description{
An internal function to compute next m based on dicrete-time model
}
