% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{getNextY.discr}
\alias{getNextY.discr}
\title{An internal function to compute the next value of physiological variable Y
based on discrete-time model (Akushevich et. al., 2005)}
\usage{
getNextY.discr(y1, u, R, Sigma)
}
\arguments{
\item{y1}{y1}

\item{u}{u (see Akushevich et. al, 2005)}

\item{R}{R (see Akushevich et. al, 2005)}

\item{Sigma}{Sigma (see Akushevich et. al, 2005)}
}
\value{
y.next Next value of y
}
\description{
An internal function to compute the next value of physiological variable Y
based on discrete-time model (Akushevich et. al., 2005)
}
