% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{getPrevY.discr.m}
\alias{getPrevY.discr.m}
\title{An internal function to compute previous m based on discrete-time model}
\usage{
getPrevY.discr.m(y2, u, R)
}
\arguments{
\item{y2}{y2}

\item{u}{u}

\item{R}{R}
}
\value{
m Next value of m (see Yashin et. al, 2007)
}
\description{
An internal function to compute previous m based on discrete-time model
}
