% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdat.R
\docType{data}
\name{longdat}
\alias{longdat}
\title{This is the longitudinal dataset.}
\description{
This is the longitudinal dataset.
}
\author{
Ilya Y Zhbannikov \email{ilya.zhbannikov@duke.edu}
}
\keyword{data}
