% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTest.R
\name{LRTest}
\alias{LRTest}
\title{Likelihood-ratio test}
\usage{
LRTest(LA, L0, df = 1)
}
\arguments{
\item{LA}{Log-likelihood for alternative hyphotesis}

\item{L0}{Log-likelihood for null hyphotesis}

\item{df}{Degrees of freedom for Chi-square test}
}
\value{
p-value of LR test.
}
\description{
Likelihood-ratio test
}
