% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{m}
\alias{m}
\title{An internal function to compute m from}
\usage{
m(y, t1, t2, a, f1)
}
\arguments{
\item{y}{Current value of Y}

\item{t1}{t1}

\item{t2}{t2}

\item{a}{a (see Yashin et. al, 2007)}

\item{f1}{f1 (see Yashin et. al, 2007)}
}
\value{
m m (see Yashin et. al, 2007)
}
\description{
An internal function to compute m from
}
