% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.short.format.R
\name{make.short.format}
\alias{make.short.format}
\title{An internal function which construct short data format from a given long}
\usage{
make.short.format(
  x,
  col.id = 1,
  col.status = 2,
  col.t1 = 3,
  col.t2 = 4,
  col.cov = 5
)
}
\arguments{
\item{x}{Dataset}

\item{col.id}{Column ID index}

\item{col.status}{Column status index}

\item{col.t1}{Column t1 index}

\item{col.t2}{Column t2 index}

\item{col.cov}{Column covariates indices}
}
\value{
column index(es) in the provided dataset
}
\description{
An internal function which construct short data format from a given long
}
