% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{mu}
\alias{mu}
\title{An internal function to compute mu}
\usage{
mu(y, mu0, b, Q, theta, tt)
}
\arguments{
\item{y}{Current value of y}

\item{mu0}{mu0 (see Yashin et. al, 2007)}

\item{b}{b (see Yashin et. al, 2007)}

\item{Q}{Q (see Yashin et. al, 2007)}

\item{theta}{theta (see Yashin et. al, 2007)}

\item{tt}{t (time)}
}
\value{
mu Next value of mu
}
\description{
An internal function to compute mu
}
