% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_discr}
\alias{prepare_data_discr}
\title{Prepares discrete-time dataset.}
\usage{
prepare_data_discr(
  merged.data,
  interval,
  col.status.ind,
  col.id.ind,
  col.age.ind,
  col.age.event.ind,
  col.covar.ind,
  verbose
)
}
\arguments{
\item{merged.data}{a longitudinal study dataset.}

\item{interval}{interval between observations.}

\item{col.status.ind}{index of "status" column.}

\item{col.id.ind}{subject id column index.}

\item{col.age.ind}{index of the age column.}

\item{col.age.event.ind}{an index of the column which represents the time in which event occured.}

\item{col.covar.ind}{a set of column indexes which represent covariates.}

\item{verbose}{turns on/off verbosing output.
Filling the last cell}
}
\description{
Prepares discrete-time dataset.
}
