% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.impute.R
\name{sigma_sq}
\alias{sigma_sq}
\title{An internal function to compute sigma square analytically}
\usage{
sigma_sq(t1, t2, b)
}
\arguments{
\item{t1}{t1}

\item{t2}{t2}

\item{b}{b (see Yashin et. al, 2007)}
}
\value{
sigma_square (see Akushevich et. al, 2005)
}
\description{
An internal function to compute sigma square analytically
}
