% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_discr.R
\name{simdata_discr}
\alias{simdata_discr}
\title{Multi-dimension simulation function}
\usage{
simdata_discr(
  N = 100,
  a = -0.05,
  f1 = 80,
  Q = 2e-08,
  f = 80,
  b = 5,
  mu0 = 1e-05,
  theta = 0.08,
  ystart = 80,
  tstart = 30,
  tend = 105,
  dt = 1,
  nobs = NULL,
  format = "long"
)
}
\arguments{
\item{N}{Number of individuals}

\item{a}{A k by k matrix, which characterize the rate of the adaptive response.}

\item{f1}{A particular state, which is a deviation from the normal (or optimal). This is a vector with length of k.}

\item{Q}{A matrix k by k, which is a non-negative-definite symmetric matrix.}

\item{f}{A vector-function (with length k) of the normal (or optimal) state.}

\item{b}{A diffusion coefficient, k by k matrix.}

\item{mu0}{mortality at start period of time.}

\item{theta}{A displacement coefficient of the Gompertz function.}

\item{ystart}{A vector with length equal to number of dimensions used, defines starting values of covariates.
Default ystart = 80.}

\item{tstart}{Starting time (age). 
Can be a number (30 by default) or a vector of two numbers: c(a, b) - in this case, starting value of time 
is simulated via uniform(a,b) distribution.}

\item{tend}{A number, defines final time (105 by default).}

\item{dt}{A time step (1 by default).}

\item{nobs}{A number, defines a number of observations (lines) for an individual, NULL by default.}

\item{format}{Data format: "long" (default), "short".}
}
\value{
A table with simulated data.
}
\description{
Multi-dimension simulation function
}
\examples{
library(stpm)
data <- simdata_discr(N=100)
head(data)

}
\references{
Akushevich I., Kulminski A. and Manton K. (2005), Life tables with covariates: Dynamic model 
for Nonlinear Analysis of Longitudinal Data. Mathematical Population Studies, 12(2), pp.: 51-80.
<DOI:10.1080/08898480590932296>.
}
