% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_con_1d_g.r
\name{spm_con_1d_g}
\alias{spm_con_1d_g}
\title{Fitting a 1-D genetic SPM model with constant parameters}
\usage{
spm_con_1d_g(
  spm_data,
  gene_data,
  a = NA,
  b = NA,
  q = NA,
  f = NA,
  f1 = NA,
  mu0 = NA,
  theta = NA,
  effect = c("a"),
  lower = c(),
  upper = c(),
  control = list(xtol_rel = 1e-06),
  global = FALSE,
  verbose = TRUE,
  ahessian = FALSE,
  method = "lbfgs",
  method.hessian = "L-BFGS-B"
)
}
\arguments{
\item{spm_data}{A dataset for the SPM model. See the STPM pacakge for more details about the format.}

\item{gene_data}{A two column dataset containing the genotypes for the individuals in spm_data. 
The first column \code{id} is the ID of the individuals in spm_data, 
and the second column \code{geno} is the genotype.}

\item{a}{The initial value for the paramter \eqn{a}. 
The initial value will be predicted if not specified.}

\item{b}{The initial value for the paramter \eqn{b}. 
The initial value will be predicted if not specified.}

\item{q}{The initial value for the paramter \eqn{q}. 
The initial value will be predicted if not specified.}

\item{f}{The initial value for the paramter \eqn{f}. 
The initial value will be predicted if not specified.}

\item{f1}{The initial value for the paramter \eqn{f_1}. 
The initial value will be predicted if not specified.}

\item{mu0}{The initial value for the paramter \eqn{\mu_0} 
in the baseline hazard.
The initial value will be predicted if not specified.}

\item{theta}{The initial value for the paramter \eqn{\theta} 
in the baseline hazard. The initial value will be predicted if not specified.}

\item{effect}{A character vector of the parameters that are linked to genotypes. 
The vector can contain any combination of \code{a}, \code{b}, \code{q}, \code{f}, \code{mu0}.}

\item{lower}{A vector of the lower bound of the parameters.}

\item{upper}{A vector of the upper bound of the parameters.}

\item{control}{A list of the control parameters for the optimization paramters.}

\item{global}{A logical variable indicating whether the MLSL (TRUE) or 
the L-BFGS (FALSE) algorithm is used for the optimization.}

\item{verbose}{A logical variable indicating whether initial 
information is printed.}

\item{ahessian}{A logical variable indicating whether the 
approximate (FALSE) or analytical (TRUE) Hessian is returned.}

\item{method}{Optimization method. 
Can be one of the following: lbfgs, mlsl, mma, slsqp, tnewton, varmetric.
Default: \code{lbfgs.}}

\item{method.hessian}{Optimization method for hessian calculation (if ahessian=F).
Default: \code{L-BFGS-B}.}
}
\value{
est The estimates of the parameters.

hessian The Hessian matrix of the estimates.

hessian The Hessian matrix of the estimates.

lik The minus log-likelihood.

con A number indicating the convergence. See the 'nloptr' package for more details.

message Extra message about the convergence. See the 'nloptr' package for more details.

beta The coefficients of the genetic effect on the parameters to be linked to genotypes.
}
\description{
This function implements a continuous genetic SPM model by assuming all the parameters are constants.
}
\examples{
\dontrun{ 
library(stpm) 
data(ex_spmcon1dg)
res <- spm_con_1d_g(ex_data$spm_data, ex_data$gene_data, 
a = -0.02, b=0.2, q=0.01, f=3, f1=3, mu0=0.01, theta=1e-05, 
upper=c(-0.01,3,0.1,10,10,0.1,1e-05), lower=c(-1,0.01,0.00001,1,1,0.001,1e-05), 
effect=c('q'))
}
}
\references{
He, L., Zhbannikov, I., Arbeev, K. G., Yashin, A. I., and Kulminski, A.M., 2017. 
Genetic stochastic process model for detecting pleiotropic and interaction effects with longitudinal data.
}
